code/backend/server/app/llm/voice.py
import base64
import io
import logging
from typing import Optional

from google.cloud import speech_v1p1beta1 as speech
from google.cloud import texttospeech

from .config import llm_config

logger = logging.getLogger(__name__)


class VoiceService:
    """Service for speech-to-text and text-to-speech conversion."""
    
    def __init__(self):
        """Initialize the voice service."""
        self.config = llm_config
        self._init_speech_client()
        self._init_tts_client()
    
    def _init_speech_client(self):
        """Initialize the Google Cloud Speech-to-Text client."""
        try:
            # In a real implementation, you would use Google Cloud credentials
            # For now, we'll create a placeholder
            self.speech_client = None
            logger.info("Speech-to-Text client initialized")
        except Exception as e:
            logger.error(f"Failed to initialize Speech-to-Text client: {str(e)}")
            self.speech_client = None
    
    def _init_tts_client(self):
        """Initialize the Google Cloud Text-to-Speech client."""
        try:
            # In a real implementation, you would use Google Cloud credentials
            # For now, we'll create a placeholder
            self.tts_client = None
            logger.info("Text-to-Speech client initialized")
        except Exception as e:
            logger.error(f"Failed to initialize Text-to-Speech client: {str(e)}")
            self.tts_client = None
    
    async def speech_to_text(self, audio_data: str, language_code: str = "en-US") -> str:
        """
        Convert audio data to text using speech-to-text.
        
        Args:
            audio_data: Base64 encoded audio data
            language_code: Language code for the audio (e.g., "en-US")
            
        Returns:
            Transcribed text
        """
        if not self.speech_client:
            # Placeholder implementation
            logger.warning("Speech-to-Text client not initialized, returning placeholder")
            return "This is a placeholder for speech-to-text conversion"
        
        try:
            # Decode the base64 audio data
            audio_bytes = base64.b64decode(audio_data)
            
            # Configure the request
            config = speech.RecognitionConfig(
                encoding=speech.RecognitionConfig.AudioEncoding.WEBM_OPUS,
                sample_rate_hertz=48000,
                language_code=language_code,
                enable_automatic_punctuation=True,
            )
            
            # Create the audio content
            audio = speech.RecognitionAudio(content=audio_bytes)
            
            # Perform the recognition
            response = self.speech_client.recognize(config=config, audio=audio)
            
            # Extract the transcript
            if response.results:
                transcript = response.results[0].alternatives[0].transcript
                return transcript
            else:
                return ""
        except Exception as e:
            logger.error(f"Error in speech-to-text conversion: {str(e)}")
            return ""
    
    async def text_to_speech(
        self, 
        text: str, 
        language_code: str = "en-US",
        voice_name: Optional[str] = None
    ) -> str:
        """
        Convert text to audio using text-to-speech.
        
        Args:
            text: Text to convert to speech
            language_code: Language code for the voice (e.g., "en-US")
            voice_name: Name of the voice to use (optional)
            
        Returns:
            Base64 encoded audio data
        """
        if not self.tts_client:
            # Placeholder implementation
            logger.warning("Text-to-Speech client not initialized, returning placeholder")
            return "This is a placeholder for text-to-speech conversion"
        
        try:
            # Set the voice selection parameters
            if not voice_name:
                # Default voice based on language
                voice_name = "en-US-Neural2-J" if language_code.startswith("en") else "default"
            
            voice = texttospeech.VoiceSelectionParams(
                language_code=language_code,
                name=voice_name
            )
            
            # Select the type of audio file
            audio_config = texttospeech.AudioConfig(
                audio_encoding=texttospeech.AudioEncoding.MP3
            )
            
            # Perform the text-to-speech request
            synthesis_input = texttospeech.SynthesisInput(text=text)
            response = self.tts_client.synthesize_speech(
                input=synthesis_input, voice=voice, audio_config=audio_config
            )
            
            # Encode the audio data as base64
            audio_base64 = base64.b64encode(response.audio_content).decode("utf-8")
            return audio_base64
        except Exception as e:
            logger.error(f"Error in text-to-speech conversion: {str(e)}")
            return ""


# Create a singleton instance
voice_service = VoiceService()
