code/backend/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
google-generativeai==0.8.3
google-cloud-speech==2.26.1
google-cloud-texttospeech==2.16.3
